+++ src/linkittydo-web/src/App.tsx
@@ -6,6 +6,7 @@ import { UserModal } from './components/UserModal'
 import { UserManageModal } from './components/UserManageModal'
 import { HomePage } from './pages/HomePage'
 import { GameHistoryPage } from './pages/GameHistoryPage'
+import { LeaderboardPage } from './pages/LeaderboardPage'
 import { useUser } from './hooks/useUser'
 import './App.css'
 
@@ -47,6 +48,7 @@ function App() {
         <Route path="/" element={<HomePage />} />
         <Route path="/play" element={<GameBoard />} />
         <Route path="/history" element={<GameHistoryPage />} />
+        <Route path="/leaderboard" element={<LeaderboardPage />} />
       </Routes>
 
       <UserModal